requirements.txt
nonebot-adapter-qq==1.5.0
nonebot2==2.3.2
aiofiles==24.1.0
ujson==5.10.0